refactor(shell): Phase 1——project-overview.vue 抽出三个方法组模块
Phase 0 把 3906 行样式外置到 .scss 之后，本次继续拆 script：把三组
内聚、无跨组耦合的方法搬进独立 .js 模块，在 methods 里展开合并。

- panelSwitching.js（87 行，3 个方法）：左栏面板切换
- agentClientActions.js（306 行，7 个方法）：AI 指令路由
- librePool.js（111 行，10 个方法）：LOWA 保活池与活跃指针

纯搬移，零逻辑改写：方法体逐行比对 482 行无差异，this 仍指向页面实例。
project-overview.vue 6731 → 6254 行。

验证：
- 方法名扫描：三模块 20 个方法与 .vue 剩余方法零重名（无静默覆盖）
- data-v 硬编码扫描：0 命中，scope id 变动不影响任何选择器
- npm run build:h5 通过；npm run check:emits 通过
- npm run test:app-e2e 29/29 通过

顺带在 <style> 上方标注样式单一来源为 project-overview.scss。
