+++ frontend/src/pages/project-overview/agentClientActions.js
@@ -1,0 +1,306 @@
+// project-overview.vue 的 AI 指令路由：SSE client_action 分发（含 doc_*/wps_* 双轨去重）、
+// doc 流式写入缓冲、编辑器打开/重载/命令执行与结果回传。
+// 经展开进组件 methods（纯搬移，Phase 1 外置），`this` 即 project-overview 页面实例。
+import { sendEditorResult, getFileDetail } from '@/services/api.js'
+
+export const agentClientActionMethods = {
+    handleClientAction(action) {
+        console.log('[ProjectOverview] Client Action:', action)
+
+        // 双轨迁移（docs/AI_ARCHITECTURE.md Phase 3）：新后端对每条指令按"新名在前、旧名在后"
+        // 各发一份。一旦见到任一新名即判定为新后端（SSE 单连接有序），此后丢弃所有旧名事件，
+        // 保证新旧后端搭配下每条指令都恰好执行一次。
+        const isNewName = action.tool === 'editor_command' ||
+            ['doc_open_file', 'doc_reload_file', 'doc_stream_data'].includes(action.action)
+        const isLegacyName = action.tool === 'wps_command' ||
+            ['wps_open_file', 'wps_reload_file', 'wps_stream_data'].includes(action.action)
+        if (isNewName) this._editorContractV2 = true
+        if (isLegacyName && this._editorContractV2) return
+
+        if (action.action === 'refresh_files') {
+            if (this.$refs.fileTree && this.$refs.fileTree.loadFiles) {
+                console.log('[ProjectOverview] Refreshing File Tree...')
+                this.$refs.fileTree.loadFiles()
+                uni.showToast({ title: '文件已更新', icon: 'none' })
+            }
+        }
+        // AI Agent 请求打开文件
+        else if (action.action === 'doc_open_file' || action.action === 'wps_open_file') {
+            this.handleEditorOpenFile(action)
+        }
+        // AI Agent 请求重新加载文件（用于后端修改文件后刷新编辑器）
+        else if (action.action === 'doc_reload_file' || action.action === 'wps_reload_file') {
+            this.handleEditorReloadFile(action)
+        }
+        // AI Agent 请求执行编辑器命令
+        else if (action.tool === 'editor_command' || action.tool === 'wps_command') {
+            // 特殊处理同步打开命令（新建文件流式写入）
+            if (action.action === 'doc_open_file_sync' || action.action === 'wps_open_file_sync') {
+                this.handleEditorOpenFileSync(action)
+            } else {
+                this.handleEditorCommand(action)
+            }
+        }
+        // 后端流式写入数据（doc_start_stream 工具）：缓冲后经 LibreOffice 执行器落字
+        else if (action.action === 'doc_stream_data' || action.action === 'wps_stream_data') {
+            this.handleDocStreamData(action.content || '')
+        }
+    },
+
+    // --- 流式写入（#79：LibreOffice 消费端，替代原 useWpsBridge.handleWpsStreamData）---
+    // 与原 WPS 实现同构：本地缓冲 + 定时批量 flush，减少 worker 往返。
+    handleDocStreamData(content) {
+        if (!content) return
+        this._docStreamBuffer = (this._docStreamBuffer || '') + content
+        if (!this._docStreamTimer) {
+            this._docStreamTimer = setTimeout(() => {
+                this._docStreamTimer = null
+                this.flushDocStreamBuffer()
+            }, 150)
+        }
+    },
+    async flushDocStreamBuffer() {
+        if (!this._docStreamBuffer || this._docStreamBusy) return
+        if (!this.libreOfficeActive || !this.libreOfficeExecutor) return
+        this._docStreamBusy = true
+        const text = this._docStreamBuffer
+        this._docStreamBuffer = ''
+        try {
+            await this.libreOfficeExecutor.executeCommand('insert_at_cursor', { text })
+        } catch (e) {
+            console.error('[ProjectOverview] doc stream insert error:', e)
+        } finally {
+            this._docStreamBusy = false
+            if (this._docStreamBuffer && !this._docStreamTimer) {
+                this._docStreamTimer = setTimeout(() => {
+                    this._docStreamTimer = null
+                    this.flushDocStreamBuffer()
+                }, 150)
+            }
+        }
+    },
+
+    /**
+     * 处理 AI Agent 的同步打开文件请求 (用于流式写入)
+     * 打开文件，等待内置 LibreOffice 编辑器就绪后返回结果给后端（#79）
+     */
+    async handleEditorOpenFileSync(action) {
+        console.log('[ProjectOverview] Open File Sync:', action)
+        const { params, requestId, conversationId } = action
+
+        try {
+            if (!params || !params.fileId) {
+                console.error('[ProjectOverview] No fileId in doc_open_file_sync')
+                await sendEditorResult(conversationId, requestId, false, null, '缺少文件ID')
+                return
+            }
+
+            // 1. 刷新文件列表以获取最新文件
+            if (this.$refs.fileTree && this.$refs.fileTree.loadFiles) {
+                await this.$refs.fileTree.loadFiles()
+            }
+
+            // 2. 获取文件详情
+            const file = await getFileDetail(this.projectId, params.fileId)
+            if (!file) {
+                console.error('[ProjectOverview] File not found:', params.fileId)
+                await sendEditorResult(conversationId, requestId, false, null, '文件不存在')
+                return
+            }
+
+            console.log('[ProjectOverview] Opening file for streaming:', file.name)
+
+            // 3. 打开文件（挂载/激活内置 LibreOffice 编辑器）
+            await this.openFile(file)
+
+            // 4. 等待编辑器就绪（onLibreReady 置位，最多等待 90 秒——LOWA 首次 boot 较慢）
+            let editorReady = false
+            for (let i = 0; i < 180; i++) {
+                await new Promise(resolve => setTimeout(resolve, 500))
+                if (this.libreOfficeActive && this.libreOfficeExecutor) {
+                    editorReady = true
+                    console.log('[ProjectOverview] LibreOffice editor ready after', (i + 1) * 500, 'ms')
+                    break
+                }
+            }
+
+            if (!editorReady) {
+                console.error('[ProjectOverview] LibreOffice editor not ready after timeout')
+                await sendEditorResult(conversationId, requestId, false, null, '编辑器未就绪')
+                return
+            }
+
+            // 5. 重置流式缓冲，准备接收新的流式数据
+            this._docStreamBuffer = ''
+            if (this._docStreamTimer) { clearTimeout(this._docStreamTimer); this._docStreamTimer = null }
+            this._docStreamBusy = false
+            console.log('[ProjectOverview] Stream state reset, ready for streaming')
+
+            // 6. 返回成功给后端
+            console.log('[ProjectOverview] Open File Sync success')
+            await sendEditorResult(conversationId, requestId, true, {
+                fileId: file.id,
+                fileName: file.name,
+                status: 'ready'
+            }, null)
+
+            uni.showToast({ title: `已打开: ${file.name}`, icon: 'none' })
+
+        } catch (e) {
+            console.error('[ProjectOverview] handleEditorOpenFileSync error:', e)
+            await sendEditorResult(conversationId, requestId, false, null, e.message)
+        }
+    },
+
+    /**
+     * 处理 AI Agent 的打开文件请求
+     */
+    async handleEditorOpenFile(action) {
+        console.log('[ProjectOverview] WPS Open File:', action)
+        try {
+            const fileId = action.fileId
+            if (!fileId) {
+                console.warn('[ProjectOverview] No fileId in doc_open_file action')
+                return
+            }
+
+            // 获取文件详情
+            const file = await getFileDetail(this.projectId, fileId)
+            if (!file) {
+                console.error('[ProjectOverview] File not found:', fileId)
+                uni.showToast({ title: '文件不存在', icon: 'none' })
+                return
+            }
+
+            // 打开文件
+            this.openFile(file)
+
+            // 提示用户
+            uni.showToast({ title: `已打开: ${file.name}`, icon: 'none' })
+
+        } catch (e) {
+            console.error('[ProjectOverview] handleEditorOpenFile error:', e)
+            uni.showToast({ title: '打开文件失败', icon: 'none' })
+        }
+    },
+
+    /**
+     * 处理 AI Agent 的重新加载文件请求
+     * 当后端修改了文件后，需要通知前端刷新编辑器以显示最新内容
+     *
+     * 工作原理：
+     * 1. 后端修改文件后会更新 wpsFileId（通用文件 ID，添加版本时间戳）
+     * 2. 前端获取最新文件信息，更新 leftFiles/rightFiles 中的 wpsFileId
+     * 3. LibreOfficeEditor 以文件为 key/prop，检测到变化后重新加载
+     */
+    async handleEditorReloadFile(action) {
+        console.log('[ProjectOverview] WPS Reload File:', action)
+        try {
+            const fileId = action.fileId
+            if (!fileId) {
+                console.warn('[ProjectOverview] No fileId in doc_reload_file action')
+                return
+            }
+
+            // 获取文件详情（确保获取最新信息，包括新的 wpsFileId）
+            const file = await getFileDetail(this.projectId, fileId)
+            if (!file) {
+                console.error('[ProjectOverview] File not found:', fileId)
+                uni.showToast({ title: '文件不存在', icon: 'none' })
+                return
+            }
+
+            console.log('[ProjectOverview] Got updated file info:', {
+                id: file.id,
+                name: file.name,
+                wpsFileId: file.wpsFileId
+            })
+
+            // 同时更新 leftFiles 和 rightFiles 中的文件信息
+            // 这样可以确保所有打开的相同文件都能获得新的 wpsFileId
+            let updated = false
+
+            // 更新左侧窗格
+            const existingLeft = this.leftFiles.find(f => f.id === file.id)
+            if (existingLeft) {
+                const oldWpsFileId = existingLeft.wpsFileId
+                Object.assign(existingLeft, file)
+                console.log('[ProjectOverview] Updated leftFiles:', {
+                    oldWpsFileId,
+                    newWpsFileId: file.wpsFileId
+                })
+                updated = true
+            }
+
+            // 更新右侧窗格
+            const existingRight = this.rightFiles.find(f => f.id === file.id)
+            if (existingRight) {
+                const oldWpsFileId = existingRight.wpsFileId
+                Object.assign(existingRight, file)
+                console.log('[ProjectOverview] Updated rightFiles:', {
+                    oldWpsFileId,
+                    newWpsFileId: file.wpsFileId
+                })
+                updated = true
+            }
+
+            // 保活池实例不再"切回即重挂载"：后台常驻实例里还是旧内容。把该
+            // 文件的非活动保活实例逐出 LRU（卸载），下次激活时重挂载并拉取
+            // 新 wpsFileId 的字节。当前正显示的实例保持改前行为（不强刷）。
+            if (updated) {
+                this.libreLruKeys = this.libreLruKeys.filter(k => {
+                    if (!k.endsWith(':' + file.id)) return true
+                    return k === 'left:' + this.activeFileIdLeft || k === 'right:' + this.activeFileIdRight
+                })
+            }
+
+            // 如果文件不在任何窗格中打开，则打开它
+            if (!updated) {
+                console.log('[ProjectOverview] File not open, opening:', file.name)
+                this.openFile(file)
+            }
+
+            uni.showToast({ title: `文件已更新: ${file.name}`, icon: 'success' })
+
+            // 刷新文件树以更新文件信息
+            if (this.$refs.fileTree && this.$refs.fileTree.loadFiles) {
+                this.$refs.fileTree.loadFiles()
+            }
+
+        } catch (e) {
+            console.error('[ProjectOverview] handleEditorReloadFile error:', e)
+            uni.showToast({ title: '刷新文件失败', icon: 'none' })
+        }
+    },
+
+    /**
+     * 处理 AI Agent 的编辑器命令请求（#79：LibreOffice 是唯一执行器；
+     * 结果经 sendEditorResult 回传后端，路由 /editor-result，双轨迁移见 Phase 3）
+     */
+    async handleEditorCommand(action) {
+        console.log('[ProjectOverview] ========== Editor Command Start ==========')
+        console.log('[ProjectOverview] Editor Command:', JSON.stringify(action))
+
+        const { action: commandAction, params, requestId, conversationId } = action
+        console.log('[ProjectOverview] commandAction:', commandAction, 'requestId:', requestId)
+
+        if (!this.libreOfficeActive || !this.libreOfficeExecutor) {
+            console.error('[ProjectOverview] No embedded editor available')
+            await sendEditorResult(conversationId, requestId, false, null, '编辑器未就绪，请先打开一个文档')
+            return
+        }
+
+        try {
+            // __agent 标记：worker 据此把这条命令产生的修订署名为 AI Workdeck
+            //（用户本人的 IME 输入等不带标记，署用户名），修订面板里可区分来源。
+            const result = await this.libreOfficeExecutor.executeCommand(
+                commandAction, Object.assign({}, params, { __agent: true }))
+            const successFlag = result && result.success !== false
+            await sendEditorResult(conversationId, requestId, successFlag, result, (result && result.error) || null)
+        } catch (e) {
+            console.error('[ProjectOverview] LibreOffice command error:', e)
+            await sendEditorResult(conversationId, requestId, false, null, e.message)
+        }
+        console.log('[ProjectOverview] ========== Editor Command End ==========')
+    },
+}
